requirements.txt
pytesseract
python-dotenv
requests
numpy
pdf2image
Flask
Flask_Cors